owner for edit/delete

+++ app/Http/Controllers/ListingController.php
@@ -62,6 +62,11 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing) {
         //dd($request->file('logo'));
 
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->user()) {
+            abort(403, 'Unauthprized');
+        }
+
         $formFields = $request->validate([
             'company' => ['required'],
             'title' => 'required',
@@ -83,6 +88,11 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->user()) {
+            abort(403, 'Unauthprized');
+        }
+        
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully');
